+++ resources/views/admin/dashboard.blade.php
@@ -1,8 +1,0 @@
-@extends('gotime::layouts.admin')
-
-@section('content')
-
-<h1>Admin Dashboard</h1>
-
-{{-- <x-gotime-menu menuname="admin"></x-gotime-menu> --}}
-@endsection