+++ resources/views/components/auth-box.blade.php
@@ -1,6 +1,6 @@
 {{-- auth template layout --}}
 <div class="py-res px">
-    <div class="mw600 max bx">
+    <div class="bx maxw600 max">
         {{ $slot }}
     </div>
 </div>