+++ resources/views/components/user-navigation.blade.php
@@ -1,0 +1,5 @@
+<x-gotime-menu menuname="main" filename="nav-user" class="menu">
+
+    <x-authit::logout-link />
+
+</x-gotime-menu>