+++ stubs/Providers/FortifyServiceProvider.php
@@ -37,7 +37,7 @@ class FortifyServiceProvider extends ServiceProvider
         Fortify::resetUserPasswordsUsing(ResetUserPassword::class);
 
         RateLimiter::for('login', function (Request $request) {
-            return Limit::perMinute(5)->by($request->email.$request->ip());
+            return Limit::perMinute(5)->by($request->email . $request->ip());
         });
 
         RateLimiter::for('two-factor', function (Request $request) {
@@ -59,7 +59,7 @@ class FortifyServiceProvider extends ServiceProvider
         Fortify::resetPasswordView(function ($request) {
             return view('authit::auth.reset-password', ['request' => $request]);
         });
-        
+
         Fortify::verifyEmailView(function () {
             return view('authit::auth.verify-email');
         });